+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- dial: respect "service not supported" SRV records and do not attempt to dial
+  fallback records if the server has indicated that they do not support a
+  specific service.
+
 
 ## v0.22.0 — 2024-09-23